Jupyter_Notebook + Deployment/requirements.txt
absl-py==0.12.0
apache-beam==2.29.0
astunparse==1.6.3
attrs==20.3.0
avro-python3==1.9.2.1
backcall==0.2.0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
contextlib2==0.6.0.post1
crcmod==1.7
cycler==0.10.0
Cython==0.29.23
dataclasses==0.6
decorator==5.0.7
dill==0.3.1.1
dm-tree==0.1.6
docopt==0.6.2
fastavro==1.4.0
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gin-config==0.4.0
google-api-core==1.26.3
google-api-python-client==2.2.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud-bigquery==2.14.0
google-cloud-core==1.6.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.32.0
h5py==2.10.0
hdfs==2.6.0
httplib2==0.17.4
idna==2.10
importlib-metadata==4.0.1
importlib-resources==5.1.2
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
kaggle==1.5.12
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lvis==0.5.3
lxml==4.6.3
Markdown==3.3.4
matplotlib==3.2.0
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
object-detection @ file:///D:/sign_final/inference_model/Tensorflow/models/research
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
promise==2.3
prompt-toolkit==3.0.18
proto-plus==1.18.1
protobuf==3.15.8
psutil==5.8.0
py-cpuinfo==8.0.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
pydot==1.4.2
Pygments==2.8.1
pymongo==3.11.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2021.1
pywin32==300
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.2
sentencepiece==0.1.95
seqeval==1.2.2
six==1.15.0
-e git+https://github.com/tensorflow/models@120953f59af06463a3712eb89d80a061bf269289#egg=slim&subdirectory=research\slim
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-addons==0.12.1
tensorflow-datasets==4.2.0
tensorflow-estimator==2.4.0
tensorflow-hub==0.12.0
tensorflow-metadata==0.30.0
tensorflow-model-optimization==0.5.0
termcolor==1.1.0
text-unidecode==1.3
tf-models-official==2.4.0
tf-slim==1.1.0
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
typeguard==2.12.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==1.0.1
wget==3.2
wrapt==1.12.1
zipp==3.4.1
